Changed pool to mobc pool for AMQP notifier.

+++ src/config.rs
@@ -168,53 +168,43 @@ pub struct InfoStoreOptions {
 
 pub struct AMQPHooksOptions {
     /// Url for AMQP server.
-    #[cfg(feature = "amqp_notifier")]
     #[arg(long, env = "RUSTUS_HOOKS_AMQP_URL")]
     pub hooks_amqp_url: Option<String>,
 
     /// Rustus will create exchange if enabled.
-    #[cfg(feature = "amqp_notifier")]
     #[arg(long, env = "RUSTUS_HOOKS_AMQP_DECLARE_EXCHANGE")]
     pub hooks_amqp_declare_exchange: bool,
 
     /// Rustus will create all queues for communication and bind them
     /// to exchange if enabled.
-    #[cfg(feature = "amqp_notifier")]
     #[arg(long, env = "RUSTUS_HOOKS_AMQP_DECLARE_QUEUES")]
     pub hooks_amqp_declare_queues: bool,
 
     /// Durability type of exchange.
-    #[cfg(feature = "amqp_notifier")]
     #[arg(long, env = "RUSTUS_HOOKS_AMQP_DURABLE_EXCHANGE")]
     pub hooks_amqp_durable_exchange: bool,
 
     /// Durability type of queues.
-    #[cfg(feature = "amqp_notifier")]
     #[arg(long, env = "RUSTUS_HOOKS_AMQP_DURABLE_QUEUES")]
     pub hooks_amqp_durable_queues: bool,
 
     /// Adds celery specific headers.
-    #[cfg(feature = "amqp_notifier")]
     #[arg(long, env = "RUSTUS_HOOKS_AMQP_CELERY")]
     pub hooks_amqp_celery: bool,
 
     /// Name of amqp exchange.
-    #[cfg(feature = "amqp_notifier")]
     #[arg(long, env = "RUSTUS_HOOKS_AMQP_EXCHANGE", default_value = "rustus")]
     pub hooks_amqp_exchange: String,
 
     /// Exchange kind.
-    #[cfg(feature = "amqp_notifier")]
     #[arg(long, env = "RUSTUS_HOOKS_AMQP_EXCHANGE_KIND", default_value = "topic")]
     pub hooks_amqp_exchange_kind: String,
 
     /// Routing key to use when sending message to an exchange.
-    #[cfg(feature = "amqp_notifier")]
     #[arg(long, env = "RUSTUS_HOOKS_AMQP_ROUTING_KEY")]
     pub hooks_amqp_routing_key: Option<String>,
 
     /// Prefix for all AMQP queues.
-    #[cfg(feature = "amqp_notifier")]
     #[arg(
         long,
         env = "RUSTUS_HOOKS_AMQP_QUEUES_PREFIX",
@@ -223,30 +213,26 @@ pub struct AMQPHooksOptions {
     pub hooks_amqp_queues_prefix: String,
 
     /// Maximum number of connections for RabbitMQ.
-    #[cfg(feature = "amqp_notifier")]
     #[arg(
         long,
         env = "RUSTUS_HOOKS_AMQP_CONNECTION_POOL_SIZE",
         default_value = "10"
     )]
-    pub hooks_amqp_connection_pool_size: u32,
+    pub hooks_amqp_connection_pool_size: u64,
 
     /// Maximum number of opened channels for each connection.
-    #[cfg(feature = "amqp_notifier")]
     #[arg(
         long,
         env = "RUSTUS_HOOKS_AMQP_CHANNEL_POOL_SIZE",
         default_value = "10"
     )]
-    pub hooks_amqp_channel_pool_size: u32,
+    pub hooks_amqp_channel_pool_size: u64,
 
     /// After this amount of time the connection will be dropped.
-    #[cfg(feature = "amqp_notifier")]
     #[arg(long, env = "RUSTUS_HOOKS_AMQP_IDLE_CONNECTION_TIMEOUT")]
     pub hooks_amqp_idle_connection_timeout: Option<u64>,
 
     /// After this amount of time in seconds, the channel will be closed.
-    #[cfg(feature = "amqp_notifier")]
     #[arg(long, env = "RUSTUS_HOOKS_AMQP_IDLE_CHANNELS_TIMEOUT")]
     pub hooks_amqp_idle_channels_timeout: Option<u64>,
 }
